Feat - Move TextInputBox, GridBox component to common component (#18)
