+++ src/component/Signup.tsx
@@ -2,10 +2,16 @@
 import React from 'react';
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-import { Banner, ButtonBox, ContainerComponent, Inner } from './emotion/component';
+import {
+  Banner,
+  ButtonBox,
+  ContainerComponent,
+  GridBox,
+  Inner,
+  TextInputBox,
+} from './emotion/component';
 import { HeadTitle } from './project/component';
 import theme from '../styles/theme';
-import { TextInputBox, GridBox } from './project/publish/component';
 
 export const Heading4 = styled.div`
   ${theme.textVariants.heading4}
