+++ src/component/project/detail/component.tsx
@@ -33,51 +33,3 @@ export const Body2 = ({ children, style }: { children: ReactNode; style?: Serial
     {children}
   </p>
 );
-export const GridBox = ({ children }: { children: ReactNode }) => {
-  return (
-    <div
-      css={css`
-        /* flex-direction: column; */
-        display: grid;
-        grid-template-columns: 1fr 1fr;
-      `}
-    >
-      {children}
-    </div>
-  );
-};
-export const TextInputBox = ({
-  type,
-  text,
-}: {
-  type: 'title' | 'subTitle' | 'select';
-  text: string;
-}) => {
-  const style = {
-    title: css`
-      font-size: 2.5rem;
-      font-weight: 700;
-      letter-spacing: -0.075rem;
-    `,
-    subTitle: css`
-      letter-spacing: -0.048rem;
-    `,
-    select: css`
-      font-size: 2rem;
-      letter-spacing: -0.6px;
-    `,
-  };
-  return (
-    <input
-      css={css`
-        background: none;
-        color: #fff;
-        &::placeholder {
-          color: #cbcbcb;
-        }
-        ${style[type]}
-      `}
-      placeholder={text}
-    ></input>
-  );
-};
